# docs/swagger/swagger.yaml
basePath: /api/v0
definitions:
...
  mill.Json:
    type: object
  pb.Block:
    properties:
      author:
        type: string
      body:
        type: string
      date:
        type: string
      id:
        type: string
      parents:
        items:
          type: string
        type: array
      target:
        type: string
      thread:
        type: string
      type:
        type: integer
      user:
        $ref: '#/definitions/pb.User'
        description: view info
        type: object
    type: object
  pb.BlockList:
    properties:
      items:
        items:
          $ref: '#/definitions/pb.Block'
        type: array
    type: object
...
  pb.CafeSessionList:
    properties:
      items:
        items:
          $ref: '#/definitions/pb.CafeSession'
        type: array
    type: object
...
  pb.ContactList:
    properties:
      items:
        items:
          $ref: '#/definitions/pb.Contact'
        type: array
    type: object
...
  pb.Keys:
    properties:
      files:
        type: object
    type: object
...
  pb.QueryResult:
    properties:
      date:
        type: string
      id:
        type: string
      local:
        type: boolean
      value:
        type: string
    type: object
...
host: 127.0.0.1:40600
info:
  contact:
    email: contact@textile.io
    name: Textile
    url: https://textile.io/
  description: Textile Node Local REST API Documentation
  license:
    name: MIT License
    url: https://github.com/textileio/go-textile/blob/master/LICENSE
  termsOfService: https://github.com/textileio/go-textile/blob/master/TERMS
  title: Textile Node API
  version: "0"
paths:
...
  /account/backups:
    post:
      description: Shows all known account peers
      parameters:
      - default: wait=5,events="false"
        description: 'wait: Stops searching after ''wait'' seconds have elapsed (max
          10s), events: Whether to emit Server-Sent Events (SSEvent) or plain JSON'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: results stream
          schema:
            $ref: '#/definitions/pb.QueryResult'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Show account peers
      tags:
      - account
  /account/peers:
    get:
      description: Shows all known account peers
      produces:
      - application/json
      responses:
        "200":
          description: peers
          schema:
            $ref: '#/definitions/pb.ContactList'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Show account peers
      tags:
      - account
  /blocks:
    get:
      description: |-
        Paginates blocks in a thread. Blocks are the raw components in a thread.
        Think of them as an append-only log of thread updates where each update is
        hash-linked to its parent(s). New / recovering peers can sync history by simply
        traversing the hash tree.
      parameters:
      - default: thread=,offset=,limit=5
        description: 'thread: Thread ID (can also use ''default''), offset: Offset
          ID to start listing from (omit for latest), limit: List page size (default:
          5)'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: blocks
          schema:
            $ref: '#/definitions/pb.BlockList'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Paginates blocks in a thread
      tags:
      - blocks
  /blocks/{id}:
    delete:
      description: Removes a thread block by ID
      parameters:
      - description: block id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: block
          schema:
            $ref: '#/definitions/pb.Block'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Remove thread block
      tags:
      - blocks
    get:
      description: Gets a thread block by ID
      parameters:
      - description: block id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: block
          schema:
            $ref: '#/definitions/pb.Block'
            type: object
        "404":
          description: Not Found
          schema:
            type: string
      summary: Gets thread block
      tags:
      - blocks
...
  /cafes:
    get:
      description: |-
        List info about all active cafe sessions. Cafes are other peers on the network
        who offer pinning, backup, and inbox services
      produces:
      - application/json
      responses:
        "200":
          description: cafe sessions
          schema:
            $ref: '#/definitions/pb.CafeSessionList'
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List info about all active cafe sessions
      tags:
      - cafes
    post:
      description: |-
        Registers with a cafe and saves an expiring service session token. An access
        token is required to register, and should be obtained separately from the target
        Cafe
      parameters:
      - description: cafe host
        in: header
        name: X-Textile-Args
        required: true
        type: string
      - default: token=
        description: 'token: An access token supplied by the Cafe'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: cafe session
          schema:
            $ref: '#/definitions/pb.CafeSession'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register with a Cafe
      tags:
      - cafes
  /cafes/{id}:
    delete:
      description: Deregisters with a cafe (content will expire based on the cafe's
        service rules)
      parameters:
      - description: cafe id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Deregisters a cafe
      tags:
      - cafes
    get:
      description: |-
        Gets and displays info about a cafe session. Cafes are other peers on the network
        who offer pinning, backup, and inbox services
      parameters:
      - description: cafe id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: cafe session
          schema:
            $ref: '#/definitions/pb.CafeSession'
            type: object
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Gets and displays info about a cafe session
      tags:
      - cafes
  /cafes/messages:
    post:
      description: |-
        Check for messages at all cafes. New messages are downloaded and processed
        opportunistically.
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Check for messages at all cafes
      tags:
      - cafes
  /config:
    patch:
      consumes:
      - application/json
      description: |-
        When patching config values, valid JSON types must be used. For example, a string
        should be escaped or wrapped in single quotes (e.g., \"127.0.0.1:40600\") and
        arrays and objects work fine (e.g. '{"API": "127.0.0.1:40600"}') but should be
        wrapped in single quotes. Be sure to restart the daemon for changes to take effect.
        See https://tools.ietf.org/html/rfc6902 for details on RFC6902 JSON patch format.
      parameters:
      - description: An RFC6902 JSON patch (array of ops)
        in: body
        name: patch
        required: true
        schema:
          $ref: '#/definitions/mill.Json'
          type: object
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get active config settings
      tags:
      - config
    put:
      consumes:
      - application/json
      description: |-
        Replace entire config file contents. The config command controls configuration
        variables. It works much like 'git config'. The configuration values are stored
        in a config file inside the Textile repository.
      parameters:
      - description: JSON document
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/mill.Json'
          type: object
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Replce config settings.
      tags:
      - config
  /config/{path}:
    get:
      description: |-
        Report the currently active config settings, which may differ from the values
        specifed when setting/patching values.
      parameters:
      - description: config path (e.g., Addresses/API)
        in: path
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: new config value
          schema:
            $ref: '#/definitions/mill.Json'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get active config settings
      tags:
      - config
  /contacts:
    get:
      description: Lists all, or thread-based, contacts.
      parameters:
      - default: thread=
        description: 'thread: Thread ID (omit for all known contacts)'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: contacts
          schema:
            $ref: '#/definitions/pb.ContactList'
            type: object
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List known contacts
      tags:
      - contacts
  /contacts/{id}:
    delete:
      description: Removes a known contact
      parameters:
      - description: contact id
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: ok
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Remove a contact
      tags:
      - contacts
    get:
      description: Gets information about a known contact
      parameters:
      - description: contact id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: contact
          schema:
            $ref: '#/definitions/pb.Contact'
            type: object
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get a known contact
      tags:
      - contacts
    put:
      consumes:
      - application/json
      description: |-
        Adds a contact to list of known local contacts. A common workflow is to pipe
        /contact/search results into this endpoint, just be sure you know what the results
        of the search are before adding!
      parameters:
      - description: contact id
        in: path
        name: id
        required: true
        type: string
      - description: contact
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/pb.Contact'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Add to known contacts
      tags:
      - contacts
  /contacts/search:
    post:
      description: Search for contacts known locally and on the network
      parameters:
      - default: local="false",limit=5,wait=5,peer=,address=,username=,events="false"
        description: 'local: Whether to only search local contacts, limit: Stops searching
          after limit results are found, wait: Stops searching after ''wait'' seconds
          have elapsed (max 10s), username: search by username string, peer: search
          by peer id string, address: search by account address string, events: Whether
          to emit Server-Sent Events (SSEvent) or plain JSON'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: results stream
          schema:
            $ref: '#/definitions/pb.QueryResult'
            type: object
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Search for contacts
      tags:
      - contacts
  /feed:
    get:
      description: |-
        Paginates post (join|leave|files|message) and annotation (comment|like) block types
        The mode option dictates how the feed is displayed:
        "chrono": All feed block types are shown. Annotations always nest their target post,
        i.e., the post a comment is about.
        "annotated": Annotations are nested under post targets, but are not shown in the
        top-level feed.
        "stacks": Related blocks are chronologically grouped into "stacks". A new stack is
        started if an unrelated block breaks continuity. This mode is used by Textile
        Photos. Stacks may include:
        * The initial post with some nested annotations. Newer annotations may have already
        been listed.
        * One or more annotations about a post. The newest annotation assumes the "top"
        position in the stack. Additional annotations are nested under the target.
        Newer annotations may have already been listed in the case as well.
      parameters:
      - default: thread=,offset=,limit=5,mode="chrono"
        description: 'thread: Thread ID (can also use ''default''), offset: Offset
          ID to start listing from (omit for latest), limit: List page size (default:
          5), mode: Feed mode (one of ''chrono'', ''annotated'', or ''stacks'')'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: feed
          schema:
            $ref: '#/definitions/pb.FeedItemList'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Paginates post and annotation block types
      tags:
      - feed
  /files:
    get:
      description: |-
        Paginates thread files. If thread id not provided, paginate all files. Specify
        "default" to use the default thread (if set).
      parameters:
      - default: thread=,offset=,limit=5
        description: 'thread: Thread ID. Omit for all, offset: Offset ID to start
          listing from. Omit for latest, limit: List page size. (default: 5)'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: files
          schema:
            $ref: '#/definitions/pb.FilesList'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Paginates thread files
      tags:
      - files
  /files/{block}:
    get:
      description: Gets a thread file by block ID
      parameters:
      - description: block id
        in: path
        name: block
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: file
          schema:
            $ref: '#/definitions/pb.Files'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get thread file
      tags:
      - files
  /invites:
    get:
      description: Lists all pending thread invites
      produces:
      - application/json
      responses:
        "200":
          description: invites
          schema:
            $ref: '#/definitions/pb.InviteViewList'
            type: object
      summary: List invites
      tags:
      - invites
    post:
      description: Creates a direct peer-to-peer or external invite to a thread
      parameters:
      - default: thread=,peer=
        description: 'thread: Thread ID (can also use ''default''), peer: Peer ID
          (omit to create an external invite)'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: invite
          schema:
            $ref: '#/definitions/pb.NewInvite'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create an invite to a thread
      tags:
      - invites
  /invites/{id}/accept:
    post:
      description: |-
        Accepts a direct peer-to-peer or external invite to a thread. Use the key option
        with an external invite
      parameters:
      - description: invite id
        in: path
        name: id
        required: true
        type: string
      - default: key=
        description: 'key: key for an external invite'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: join block
          schema:
            $ref: '#/definitions/pb.Block'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Accept a thread invite
      tags:
      - invites
  /invites/{id}/ignore:
    post:
      description: Ignores a direct peer-to-peer invite to a thread
      parameters:
      - description: invite id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Ignore a thread invite
      tags:
      - invites
  /ipfs/cat/{cid}:
    get:
      description: Displays the data behind an IPFS CID (hash)
      parameters:
      - description: ipfs/ipns cid
        in: path
        name: cid
        required: true
        type: string
      - default: key=
        description: 'key: Key to decrypt data on-the-fly'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: data
          schema:
            items:
              type: integer
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Cat IPFS data
      tags:
      - ipfs
  /ipfs/id:
    get:
      description: Displays underlying IPFS peer ID
      produces:
      - text/plain
      responses:
        "200":
          description: peer id
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get IPFS peer ID
      tags:
      - ipfs
...
  /keys/{target}:
    get:
      description: Shows file keys under the given target from an add
      parameters:
      - description: target id
        in: path
        name: blotargetck
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: keys
          schema:
            $ref: '#/definitions/pb.Keys'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Show file keys
      tags:
      - files
  /logs/{subsystem}:
    post:
      description: |-
        List or change the verbosity of one or all subsystems log output. Textile logs
        piggyback on the IPFS event logs
      parameters:
      - description: subsystem logging identifier (omit for all)
        in: path
        name: subsystem
        type: string
      - default: level=,tex-only="false"
        description: 'level: Log-level (one of: debug, info, warning, error, critical,
          or '
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: subsystems
          schema:
            $ref: '#/definitions/core.SubsystemInfo'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Access subsystem logs
      tags:
      - utils
  /messages:
    get:
      description: Paginates thread messages
      parameters:
      - default: thread=,offset=,limit=10
        description: 'thread: Thread ID (can also use ''default'', omit for all),
          offset: Offset ID to start listing from (omit for latest), limit: List page
          size (default: 5)'
        in: header
        name: X-Textile-Opts
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: messages
          schema:
            $ref: '#/definitions/pb.TextList'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Paginates thread messages
      tags:
      - messages
...
  /profile:
    get:
      description: Gets the local node's public profile
      produces:
      - application/json
      responses:
        "200":
          description: contact
          schema:
            $ref: '#/definitions/pb.Contact'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get public profile
      tags:
      - profile
...
swagger: "2.0"